// DataDiscovery/ReportGenerationFront/src/app/core/models/PaginationResponseData.ts
export interface PaginationResponseData {
    recordCount: number;
    response: PaginationResponse;
}

export interface PaginationResponse {
    content: any;
    pageable: Pageable;
    totalElements: number;
    totalPages: number;
    last: boolean;
    size: number;
    number: number;
    sort: Sort;
    numberOfElements: number;
    first: boolean;
    empty: boolean;
}

export interface Pageable {
    sort: Sort;
    offset: number;
    pageNumber: number;
    pageSize: number;
    paged: boolean;
    unpaged: boolean;
}

export interface Sort {
    sorted: boolean;
    unsorted: boolean;
    empty: boolean;
}